# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Book API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/book": {
            "post": {
                "description": "Создаёт новую книгу в системе. Принимает данные книги в формате JSON и возвращает созданную запись.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Create a new book",
                "parameters": [
                    {
                        "description": "Параметры для создания книги. Пример: {\\",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_application_http_handlers_bookshandlers.CreateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созданная книга с её уникальным ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный формат запроса или отсутствуют обязательные поля",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/book/{id}": {
            "get": {
                "description": "Возвращает данные книги по её уникальному идентификатору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Retrieve a book by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Уникальный ID книги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные книги",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Книга не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные книги по её уникальному идентификатору. Принимает новые данные книги в формате JSON.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Update a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Уникальный ID книги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления книги. Пример: {\\",
                        "name": "book",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_application_http_handlers_bookshandlers.UpdateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновлённые данные книги",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет книгу из системы по её уникальному идентификатору.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Delete a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Уникальный ID книги",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Сообщение об успешном удалении",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/books": {
            "get": {
                "description": "Возвращает список всех книг, хранящихся в системе.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "List all books",
                "responses": {
                    "200": {
                        "description": "Массив книг",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_application_http_handlers_bookshandlers.CreateBookRequest": {
            "type": "object",
            "properties": {
                "author_ids": {
                    "description": "Пример для массива убран, чтобы избежать ошибок преобразования",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "genre": {
                    "type": "string",
                    "example": "Programming"
                },
                "isbn": {
                    "type": "string",
                    "example": "1234567890"
                },
                "title": {
                    "type": "string",
                    "example": "Go Programming"
                },
                "year": {
                    "type": "integer",
                    "example": 2025
                }
            }
        },
        "internal_application_http_handlers_bookshandlers.UpdateBookRequest": {
            "type": "object",
            "properties": {
                "author_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "genre": {
                    "type": "string",
                    "example": "Programming"
                },
                "isbn": {
                    "type": "string",
                    "example": "0987654321"
                },
                "title": {
                    "type": "string",
                    "example": "Advanced Go"
                },
                "year": {
                    "type": "integer",
                    "example": 2025
                }
            }
        },
...
    }
}